# api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api": {
      "get": {
        "operationId": "example api",
        "summary": "",
        "description": "response test message",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns Hello World!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/users": {
      "get": {
        "operationId": "UserController_show",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/users/signup": {
      "post": {
        "operationId": "example signup api",
        "summary": "",
        "description": "signup api",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSignupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/users/login": {
      "post": {
        "operationId": "example login api",
        "summary": "",
        "description": "login api",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    }
  },
  "info": {
    "title": "API",
    "description": "The API Reference",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "AppResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "メッセージ",
            "example": "Hello World!"
          }
        },
        "required": [
          "message"
        ]
      },
      "UserSignupDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "name",
            "example": "test name"
          },
          "email": {
            "type": "string",
            "description": "email",
            "example": "text@example.com"
          },
          "password": {
            "type": "string",
            "description": "password",
            "example": "TODO: password hash at server side"
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ]
      },
      "UserLoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "email",
            "example": "text@example.com"
          },
          "password": {
            "type": "string",
            "description": "password",
            "example": "TODO: password hash at server side"
          }
        },
        "required": [
          "email",
          "password"
        ]
      }
    }
  }
}